refactor(fhec): extract native binary resolver
